Extract Dial and DialContextMenu from DialBox.py
